refactor: rename ContensisFields to EntryFields, define EntryMetadata and EntryMetadataClassic separately, ensure typings are exported via index.ts, tidy up imports
