+++ src/models/EntryFields.ts
