+++ src/models/EntryMetadata.ts
@@ -1,0 +1,3 @@
+export interface EntryMetadata {
+    [key: string]: string | boolean | number | null | undefined;
+}
