+++ src/models/EntrySys.ts
@@ -1,7 +1,8 @@
 import { BaseSys } from './BaseSys';
+import { EntryMetadata } from './EntryMetadata';
 
 export interface EntrySys extends BaseSys<'entry'> {
 	allUris: string[];
-	metadata: { [key: string]: any };
+	metadata: EntryMetadata;
 	properties: { [key: string]: any };
 }
