+++ src/models/index.ts
@@ -1,11 +1,16 @@
+export * from './AssetSys';
+export * from './BaseSys';
 export * from './ClientConfigFactory';
 export * from './ClientStatic';
 export * from './Config';
 export * from './ContensisClient';
 export * from './ContensisStatic';
 export * from './Entry';
+export * from './EntryFields';
 export * from './EntryGetOptions';
 export * from './EntryListOptions';
+export * from './EntryMetadata';
+export * from './EntryMetadataClassic';
 export * from './EntrySys';
 export * from './IContentTypeOperations';
 export * from './IEntryOperations';
